More applicable command template
With the command being in a dedicated file.

Now also include standard logger naming suggestions.

Fixes datalad/datalad-extension-template#27
